Add rubocop (#149)
* Add rubocop with bixby gem for standard config
* Run rubocop before running specs

+++ lib/tasks/lae.rake
@@ -1,9 +1,9 @@
+# frozen_string_literal: true
 require 'rsolr'
 require 'faraday'
 require 'yaml'
 
 namespace :lae do
-
   # TODO: we need an optimize task when we have spellcheck going; that's when
   # the spellcheck index is built (recommended, can be changed to index time but
   # takes a while and we may not want to do it every time.)
@@ -11,7 +11,7 @@ namespace :lae do
   task index: :environment do
     data = IndexEvent.get_boxes_data
     IndexEvent.delete_index
-    data.select{ |box| !box['last_mod_prod_folder'].nil? }.each do |box|
+    data.select { |box| !box['last_mod_prod_folder'].nil? }.each do |box|
       IndexEvent.index_resource(box_id: box['id'])
     end
     IndexEvent.optimize
@@ -29,11 +29,12 @@ namespace :lae do
 
     puts "Last indexing event was `#{latest_index.task}` on #{latest_index.start}"
 
-    boxes = IndexEvent.get_boxes_data.select { |box|
+    boxes = IndexEvent.get_boxes_data.select do |box|
       !box['last_mod_prod_folder'].nil?
-    }.select { |box|
+    end
+    boxes.select do |box|
       Time.parse(box['last_mod_prod_folder']['time']) > latest_index.start
-    }
+    end
     if boxes.empty?
       puts 'There is no new data to index'
     else
@@ -49,17 +50,14 @@ namespace :lae do
 
   desc 'Delete a single record from the index (supply the id: `id=12345 lae:delete_one`)'
   task delete_one: :environment do
-    if ENV['id'].nil?
-      raise 'No record id supplied'
-    else
-      IndexEvent.delete_one(ENV['id'])
-    end
+    return IndexEvent.delete_one(ENV['id']) unless ENV['id'].nil?
+    raise 'No record id supplied'
   end
 
   namespace :solr do
     desc 'Posts fixtures to Solr'
     task :index do
-      solr = RSolr.connect :url => Blacklight.connection_config[:url]
+      solr = RSolr.connect url: Blacklight.connection_config[:url]
       content = File.read('spec/fixtures/files/208_solr_docs.xml')
       solr.update(data: content, headers: { 'Content-Type' => 'text/xml' })
       solr.commit
